Fix deadlock when closing threaded data loader.
We had a race condition between the background thread exiting, the
main thread clearing the queue, and the background thread still trying
to place a new value. Instead, if the main thread has already signaled
an exit, there is no need for the background thread to keep placing
None values.