Run most of the simulation in float32, but run solver in float64.
This helps with the ill-conditioned solve and seems to fix issues with
batch size-dependent outputs at the cost of some slowdown. At least we
confine it to just the linear solve portion with the rest run in
single precision. This requires us to be more careful with data types
to avoid accidental promotion to float64 since we run JAX with double
precision enabled, so we add a bunch of manual casts.